Fixing failing unit tests

+++ rskj-core/src/test/java/co/rsk/pcc/bto/ToBase58CheckTest.java
@@ -49,7 +49,7 @@ public class ToBase58CheckTest {
 
         Assert.assertEquals(2, fn.inputs.length);
         Assert.assertEquals(SolidityType.getType("bytes").getName(), fn.inputs[0].type.getName());
-        Assert.assertEquals(SolidityType.getType("uint8").getName(), fn.inputs[1].type.getName());
+        Assert.assertEquals(SolidityType.getType("int256").getName(), fn.inputs[1].type.getName());
 
         Assert.assertEquals(1, fn.outputs.length);
         Assert.assertEquals(SolidityType.getType("string").getName(), fn.outputs[0].type.getName());
